MainPrograms(Patrol): Done
Assignee: Kyarasu
Path: MainPrograms
Description: 数か所のみ直しましたがほとんどは問題ありませんでした。connectingのUIはナイスです。
