--- Idisclosure#2/Assets/Projects/[P]MainPrograms/[P]EndGame.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c94f6a7cf89fad14aae016c6557dd1ab
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
