--- Idisclosure#2/Assets/Projects/[P]MainPrograms/[P]EndGame/Audio.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d87127b4572efeb448b6ffbf44ffc17c
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
