--- Idisclosure#2/Assets/Projects/[P]MainPrograms/[P]EndGame/Element.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a24446a6f0bbb0744b224164cbf7bcd6
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
