--- Idisclosure#2/Assets/Projects/[P]MainPrograms/[P]EndGame/Functions.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f76dcc6ace85a414590359789bfcc06d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
