--- Idisclosure#2/Assets/Projects/[P]MainPrograms/[P]EndGame/Move&Save.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3a71b46fa6e39574398a0acf08463c7a
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
