--- Idisclosure#2/Assets/Projects/[P]MainPrograms/[P]EndGame/Move&Save/MoveTo01.cs
@@ -1,6 +1,0 @@
-using UnityEngine;
-
-public class MoveTo01
-{
-    
-}
