--- Idisclosure#2/Assets/Projects/[P]MainPrograms/[P]EndGame/Move&Save/MoveTo01.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 6edb144a728bf6a4d90cfe850ab17e44
