+++ Idisclosure#2/Assets/Projects/[S]SubScenes/[S]AppScenes/[S]Terminal/Terminal.unity.meta
